+++ src/types/accounts-widget/constants.ts
@@ -1,4 +1,4 @@
 export const ACCOUNTS_WIDGET_CONSTANTS = {
   TITLE: "Cuentas",
-  ADD_TOOLTIP: "Agregar cuenta",
+  ADD_ACCOUNT_TEXT: "Agregar cuenta",
 }
