+++ src/types/accounts-widget/index.ts
@@ -1,1 +1,2 @@
 export * from "./constants"
+export * from "./types"
